feat: create api for location finding

+++ src/main/resources/db/changelog/db.changelog-1.2.sql
@@ -4,9 +4,9 @@
 CREATE TABLE IF NOT EXISTS weather.locations
 (
     id        INT PRIMARY KEY,
-    name      VARCHAR(128) NOT NULL,
-    user_id   INT          NOT NULL REFERENCES weather.users,
-    latitude  DECIMAL      NOT NULL,
-    longitude DECIMAL      NOT NULL,
+    name      VARCHAR(128)     NOT NULL,
+    user_id   INT              NOT NULL REFERENCES weather.users,
+    latitude  DOUBLE PRECISION NOT NULL,
+    longitude DOUBLE PRECISION NOT NULL,
     CONSTRAINT user_unique_location_constraint UNIQUE (user_id, latitude, longitude)
 );
